requirements.txt
langchain==0.1.14
langchain-community==0.0.20
langchain-core==0.1.44
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==1.3.0
faiss-cpu==1.7.4
altair==4
tiktoken==0.5.2